change 16bit to 8bit for cgraflow

+++ apps/hardware_benchmarks/apps/cascade/process.cpp
@@ -11,7 +11,7 @@ using namespace Halide::Runtime;
 
 int main(int argc, char **argv) {
 
-  OneInOneOut_ProcessController<uint16_t> processor("cascade",
+  OneInOneOut_ProcessController<uint8_t> processor("cascade",
                                             {
                                               {"cpu",
                                                   [&]() { cascade(processor.input, processor.output); }
@@ -23,8 +23,8 @@ int main(int argc, char **argv) {
 
                                             });
 
-  processor.input = Buffer<uint16_t>(64, 64);
-  processor.output = Buffer<uint16_t>(60, 60);
+  processor.input = Buffer<uint8_t>(64, 64);
+  processor.output = Buffer<uint8_t>(60, 60);
   
   processor.process_command(argc, argv);
